// src/types/increase.ts
export type Increases = {
    description: string;
    date: Date;
    value: number;
}

export type Expenses = {
    description: string;
    date: Date;
    value: number;
}

export type Invests = {
    description: string;
    date: Date;
    value: number;
    goal: number;
    percentage: number;
}